requirements/development.txt
-r base.txt
pytest==8.3.5
pytest-django==4.10.0
pytest-cov==6.0.0
django-debug-toolbar==4.4.6
